Add testcase for delaying TCP ports until system wab is ready
