+++ dev/com.ibm.ws.app.manager.wab.installer_fat/publish/files/configs/testWabStartDelay.xml
@@ -1,0 +1,11 @@
+<server description="wab configuration server">
+
+    <include location="../fatTestPorts.xml"/>
+    
+    <featureManager>
+		<feature>product1:product1</feature>
+		<feature>osgiConsole-1.0</feature>
+    </featureManager>
+
+    <webContainer deferServletLoad="false"/>
+</server>
